+++ liqwid-script-export/src/ScriptExport/Export.hs
@@ -6,9 +6,10 @@ import Data.Text (intercalate, unpack)
 import ScriptExport.API (runServer)
 import ScriptExport.File (runFile)
 import ScriptExport.Options (
-  Options (FileOption, ListBuilders, ServerOption),
+  Options (FileOption, ListBuilders, ServerOption, StdoutOption),
   parseOptions,
  )
+import ScriptExport.Stdout (runStdout)
 import ScriptExport.Types (Builders, toList)
 
 {- | Program entry point. It provides both file and servant api.
@@ -21,5 +22,6 @@ exportMain builders = do
   case opts of
     ServerOption so -> runServer builders so
     FileOption fo -> runFile builders fo
+    StdoutOption so -> runStdout builders so
     ListBuilders ->
       putStrLn $ unpack $ intercalate "\n" $ toList builders